MedicalChatbot/requirements.txt
flask
requests
python-dotenv
openai
Werkzeug
PyPDF2
